feat: add needed changes for RN 0.74 (#2231)
PR bringing support for RN 0.74 and bridgeless mode to the library.

+++ src/fabric/NativeSvgRenderableModule.ts
@@ -2,7 +2,7 @@
 // its needed for codegen to work
 import type { TurboModule } from 'react-native';
 import { TurboModuleRegistry } from 'react-native';
-import type { Int32, Float } from 'react-native/Libraries/Types/CodegenTypes';
+import type { Double, Float } from 'react-native/Libraries/Types/CodegenTypes';
 
 type Rect = {
   x: Float;
@@ -26,13 +26,13 @@ type Point = {
 };
 
 export interface Spec extends TurboModule {
-  isPointInFill(tag: Int32 | null, options?: Object): boolean;
-  isPointInStroke(tag: Int32 | null, options?: Object): boolean;
-  getTotalLength(tag: Int32 | null): Float;
-  getPointAtLength(tag: Int32 | null, options?: Object): Point;
-  getBBox(tag: Int32 | null, options?: Object): Rect;
-  getCTM(tag: Int32 | null): Matrix;
-  getScreenCTM(tag: Int32 | null): Matrix;
+  isPointInFill(tag: Double | null, options?: Object): boolean;
+  isPointInStroke(tag: Double | null, options?: Object): boolean;
+  getTotalLength(tag: Double | null): Float;
+  getPointAtLength(tag: Double | null, options?: Object): Point;
+  getBBox(tag: Double | null, options?: Object): Rect;
+  getCTM(tag: Double | null): Matrix;
+  getScreenCTM(tag: Double | null): Matrix;
   getRawResource(name: string): Promise<string>;
 }
